feat: :recycle: The create method from the user module has been moved to the auth module, in the auth/signup route, and now generates the token at the time of registration.

+++ api/src/modules/users/users.service.ts
@@ -3,7 +3,7 @@ import {
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
-import { CreateUserDto } from './dto/create-user.dto';
+import { SignUpDto } from './dto/signup.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { hash } from 'bcryptjs';
 import { UsersRepository } from 'src/shared/database/repositories/users.repositories';
@@ -11,55 +11,6 @@ import { UsersRepository } from 'src/shared/database/repositories/users.reposito
 @Injectable()
 export class UsersService {
   constructor(private readonly usersRepo: UsersRepository) {}
-
-  async create(createUserDto: CreateUserDto) {
-    const { name, email, password } = createUserDto;
-
-    const emailExist = await this.usersRepo.findByEmail({
-      where: { email },
-      select: { id: true },
-    });
-
-    if (emailExist) {
-      throw new ConflictException('This email already in use');
-    }
-
-    const hashedPassword = await hash(password, 12);
-
-    const user = await this.usersRepo.create({
-      data: {
-        name,
-        email,
-        password: hashedPassword,
-        categories: {
-          createMany: {
-            data: [
-              // Income
-              { name: 'Salário', icon: 'salary', type: 'INCOME' },
-              { name: 'Freelance', icon: 'freelance', type: 'INCOME' },
-              { name: 'Outro', icon: 'other', type: 'INCOME' },
-              // Expense
-              { name: 'Casa', icon: 'home', type: 'EXPENSE' },
-              { name: 'Alimentação', icon: 'food', type: 'EXPENSE' },
-              { name: 'Educação', icon: 'education', type: 'EXPENSE' },
-              { name: 'Lazer', icon: 'fun', type: 'EXPENSE' },
-              { name: 'Mercado', icon: 'grocery', type: 'EXPENSE' },
-              { name: 'Roupas', icon: 'clothes', type: 'EXPENSE' },
-              { name: 'Transporte', icon: 'transport', type: 'EXPENSE' },
-              { name: 'Viagem', icon: 'travel', type: 'EXPENSE' },
-              { name: 'Outro', icon: 'other', type: 'EXPENSE' },
-            ],
-          },
-        },
-      },
-    });
-
-    return {
-      id: user.id,
-      name: user.name,
-      email: user.email,
-    };
-  }
 
   async findAll() {
     return await this.usersRepo.findAll({});
@@ -110,6 +61,16 @@ export class UsersService {
   }
 
   async remove(userId: string) {
+    const user = await this.usersRepo.findById({
+      where: {
+        id: userId,
+      },
+    });
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     await this.usersRepo.delete({
       where: {
         id: userId,
